backend/requirements.txt
Flask==2.0.1
pyaudio==0.2.11
SpeechRecognition==3.8.1
Flask-Cors
Flask-Bcrypt
Werkzeug==2.0.3
google-cloud-translate==3.9.0